+++ src/main/java/com/proyectoscvr/calculator/service/ICalculatorService.java
@@ -1,9 +1,11 @@
 package com.proyectoscvr.calculator.service;
 
+import java.math.BigDecimal;
+
 import com.proyectoscvr.calculator.api.dto.OperationDTO;
 
 public interface ICalculatorService {
 
-	Double calculate(OperationDTO operation);
+	BigDecimal calculate(OperationDTO operation);
 
 }
