+++ src/test/java/com/proyectoscvr/calculator/operations/AdditionOperationTest.java
@@ -2,6 +2,8 @@ package com.proyectoscvr.calculator.operations;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import java.math.BigDecimal;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -14,8 +16,8 @@ class AdditionOperationTest {
 	
 	@Test
 	void testAdditionExecuteMethod() {
-		Double res = additionOperation.execute(3.0, 5.0);
+		BigDecimal res = additionOperation.execute(BigDecimal.valueOf(3.0), BigDecimal.valueOf(5.0));
 		
-		assertEquals(8.0, res);
+		assertEquals(BigDecimal.valueOf(8.0), res);
 	}
 }
